refactor(text): css import paths
