+++ packages/text/src/index.ts
@@ -1,2 +1,8 @@
-export * from './react/index'
+import './css/index.css'
+import Code from './react/code'
+import Heading from './react/heading'
+import Label from './react/label'
+import List from './react/list/index'
+import P from './react/p'
+export { Code, Heading, Label, List, P }
 export * from './vars/index'
