+++ packages/text/src/react/__specs__/list.spec.tsx
@@ -4,7 +4,7 @@ import { axe } from 'jest-axe'
 import React from 'react'
 
 import * as stories from '../__stories__/list.story'
-import { List } from '../index'
+import { List } from '../../index'
 
 describe('Text', () => {
   describe('List', () => {
