+++ packages/text/src/react/heading.tsx
@@ -2,7 +2,6 @@ import React from 'react'
 import { useTheme, names } from '@pluralsight/ps-design-system-theme'
 import { ValueOf, classNames } from '@pluralsight/ps-design-system-util'
 
-import '../css/index.css'
 import * as vars from '../vars/index'
 
 const style = ({
