--- packages/text/src/react/index.ts
@@ -1,5 +1,0 @@
-export { default as Code } from './code'
-export { default as Heading } from './heading'
-export { default as Label } from './label'
-export { default as List } from './list/index'
-export { default as P } from './p'
