+++ packages/text/src/react/label.tsx
@@ -2,7 +2,6 @@ import { useTheme, names } from '@pluralsight/ps-design-system-theme'
 import { ValueOf, classNames } from '@pluralsight/ps-design-system-util'
 import React from 'react'
 
-import '../css/index.css'
 import * as vars from '../vars/index'
 
 const style = ({
